+++ padiff/hooks.py
@@ -15,7 +15,6 @@
 import contextlib
 from functools import partial
 from .report import current_torch_report, current_paddle_report
-from .stack_info import *
 from .utils import (
     for_each_grad_tensor,
     map_structure_and_replace_key,
@@ -222,3 +221,47 @@ def _torch_tensor_to_paddle_tensor(tt):
         return paddle.to_tensor(tt.detach().cpu().numpy())
     else:
         return tt
+
+
+"""
+    stack tools
+"""
+
+
+import os.path as osp
+import traceback
+
+
+def _is_system_package(filename):
+    exclude = [
+        "lib/python",
+        "/usr/local",
+        osp.dirname(paddle.__file__),
+        osp.dirname(torch.__file__),
+        osp.dirname(__file__),  # exclude padiff
+    ]
+    for pattern in exclude:
+        if pattern in filename:
+            return True
+    return False
+
+
+def extract_frame_summary():
+    """
+    extract the current call stack by traceback module.
+    gather the call information and put them into ReportItem to helper locate the error.
+
+    frame_summary:
+        line: line of the code
+        lineno: line number of the file
+        filename: file name of the stack
+        name: the function name.
+    """
+    frame_summarys = traceback.StackSummary.extract(traceback.walk_stack(None))
+    last_user_fs = None
+    for fs in frame_summarys:
+        if not _is_system_package(fs.filename):
+            last_user_fs = fs
+            break
+    assert last_user_fs is not None, "Error happend, can't return None."
+    return last_user_fs, frame_summarys
