+++ padiff/utils.py
@@ -15,7 +15,6 @@
 import os
 import sys
 import shutil
-import warnings
 from collections import namedtuple, Iterable
 from itertools import zip_longest
 
@@ -180,34 +179,6 @@ def is_sublayer(father_net, child_net):
         return False
     else:
         raise RuntimeError("father net is not Module / Layer")
-
-
-class TableView:
-    """
-    A search speedup wrapper class.
-    """
-
-    def __init__(self, data, key=None):
-        self.data = data
-        self.view = {}
-        assert callable(key), "Key must be callable with a paramter: x -> key."
-        for item in self.data:
-            if key(item) not in self.view:
-                self.view[key(item)] = [item]
-            else:
-                warnings.warn("Warning: duplicate key is found, use list + pop strategy.")
-                self.view[key(item)].append(item)
-
-    def __getitem__(self, key):
-        assert key in self.view, "{} is not found in index.".format(key)
-        ret = self.view[key].pop(0)  # pop for sorting.
-        return ret
-
-    def __len__(self):
-        return len(self.data)
-
-    def __contains__(self, key):
-        return key in self.view
 
 
 class TreeView:
